requirements.txt
streamlit>=1.28.0
google-genai>=0.5.0
markdown>=3.5.0
protobuf<5.0.0
st-gsheets-connection>=0.1.0
gspread<6.0.0
pandas>=2.0.0